added kerberos examples.
