+++ examples/demo/src/main/resources/application.conf
@@ -5,13 +5,26 @@ examples {
     user="dbadmin"
     password=""
     filepath="webhdfs://hdfs:50070/data/"
+}
 
-#     S3 settings
-    filepath-s3="s3a://test"
+s3 {
+    filepath="s3a://test"
     aws_endpoint="minio:9000"
     aws_enable_ssl="false"
     aws_enable_path_style="true"
     aws_access_key_id="minioadmin"
     aws_secret_access_key="minioadmin"
+}
+
+gcs {
+# input your GCS configurations here
+}
+
+kerberos {
+    user="user1"
+    filepath="hdfs://hdfs.example.com:8020"
+    kerberos_service_name="vertica"
+    kerberos_host_name="vertica.example.com"
+    jaas_config_name="Client"
 }
